requirements.txt
teslapy == 2.9.0
paho-mqtt == 2.1.0
pytz == 2024.1
